requirements.txt
#debug
ipdb
ipython

#framework
django==3.1.7
werkzeug==1.0.1
django-extensions==3.1.1
django-cleanup==5.1.0
django-modeltranslation==0.16.2
django-recaptcha==2.0.6

#oauth
django-cors-headers==3.7.0
django-oauth-toolkit==1.4.0

#rest
django-tastypie==0.14.3

#util
Pillow===8.1.0
requests-oauthlib===1.3.0

